reference defaults should be forbidden, not cloned
the new parser now makes all defaults into coderefs, so this code path
isn't even taken anymore. we could probably optimize things such that
non-reference constants are stored as their value directly rather than
through a coderef, but even in that case we won't want to do it for
references.