fix comparison
this is actually wrong - the original code used a peek of 8 so that it
could match against /extends\b/ (to avoid catching things that just had
extends as a prefix). this will only make an actual difference in
obscure error handling cases though (role Foo extendsblah Bar { } will
now also give an error of "roles can't use extends"), so that should be
fine for now, since we're going to clean this bit of code up later
anyway.