get rid of parse_ident, it duplicates parse_name_prefix
the trick is that myck_rv2sv_twigils may be called during string
interpolation, and lex_peek_unichar/lex_read_unichar will automatically
read in the next chunk if they get to the end of the buffer.
unfortunately, they have a bug where they read the next chunk of the
main buffer into the string parsing sublex buffer, so we have to take
extra measures to prevent this from happening.